+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
-## v1.5.0
+## v1.6.0
 
 ### Added
 
 - Trait `AppVersionAssertionsTrait`
+
+## v1.5.0
+
+### Added
+
+- `assertArrayStructure` method in `AdditionasAssertionsTrait` [#7]
+- Trait `WithMemoryClean`
 
 ## v1.4.0
 
@@ -97,3 +104,6 @@
 ## v1.0.0
 
 ### First release
+
+
+[#7]:https://github.com/avto-dev/dev-tools/issues/7